backend/requirements.txt
Authlib==1.3.2
blinker==1.7.0
boto3==1.34.2
botocore==1.34.2
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.1
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.26.2
Pillow==10.1.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
s3transfer==0.9.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
